fix(code): extract treasury balance functionality
- adapt handleApiError pattern
- move functionality
- agent instructions to clarify /lib/hooks /hooks difference
